Implements one-time registration of probes in a process
This commit adds support for guaranteed at-most-once registration of
probes, within a single process lifetime. The linker section into which
the probes are written is now writable. When reading a probe record, the
code checks that the probe's version is supported, and, if so,
atomically exchanges it for a sentinel value (255), returning the
original. Since any probe version greater than the crate's version is
skipped, probes that are handled will be skipped in any other invocation
of the `register_probes` function, as those probes' version is now the
sentinel. Unhandled probes will not be modified, but will still be
skipped, letting later versions of the library handle them.
